etl/requirements.txt
aiosqlite==0.19.0
alembic==1.12.0
anyio==3.7.1
apprise==1.5.0
asgi-lifespan==2.1.0
asyncpg==0.28.0
attrs==23.1.0
azure-core==1.29.3
azure-identity==1.14.0
azure-storage-blob==12.17.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
coolname==2.2.0
croniter==1.4.1
cryptography==41.0.3
databricks-sdk==0.8.0
databricks-sql-connector==2.9.3
dateparser==1.1.8
docker==6.1.3
et-xmlfile==1.1.0
exceptiongroup==1.1.3
fastapi==0.103.1
fsspec==2023.9.0
google-auth==2.22.0
graphviz==0.20.1
greenlet==2.0.2
griffe==0.36.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.3
httpx==0.24.1
hyperframe==6.0.1
idna==3.4
isodate==0.6.1
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kubernetes==27.2.0
lz4==4.3.2
Mako==1.2.4
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mercurial==6.1.1
msal==1.23.0
msal-extensions==1.0.0
numpy==1.25.2
oauthlib==3.2.2
openpyxl==3.1.2
orjson==3.9.5
packaging==23.1
pandas==2.1.0
pathspec==0.11.2
pendulum==2.1.2
portalocker==2.7.0
prefect==2.12.1
psycopg==3.1.10
psycopg-binary==3.1.10
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.12
pydicom==2.4.3
Pygments==2.16.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3
pytzdata==2020.1
PyYAML==6.0.1
readchar==4.0.5
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.5.2
rpds-py==0.10.2
rsa==4.9
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.49
sqlparse==0.4.4
starlette==0.27.0
text-unidecode==1.3
thrift==0.16.0
toml==0.10.2
typer==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.16
uvicorn==0.23.2
websocket-client==1.6.2
websockets==11.0.3
